Refine contribution e86984d6
* Make ReflectionUtils#getWriteMethod public
* Move chained setters tests into a dedicated class
