--- easy-random-core/src/test/java/org/jeasy/random/EasyRandomChainedSettersTest.java
@@ -1,82 +1,0 @@
-/*
- * The MIT License
- *
- *   Copyright (c) 2020, Mahmoud Ben Hassine (mahmoud.benhassine@icloud.com)
- *
- *   Permission is hereby granted, free of charge, to any person obtaining a copy
- *   of this software and associated documentation files (the "Software"), to deal
- *   in the Software without restriction, including without limitation the rights
- *   to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- *   copies of the Software, and to permit persons to whom the Software is
- *   furnished to do so, subject to the following conditions:
- *
- *   The above copyright notice and this permission notice shall be included in
- *   all copies or substantial portions of the Software.
- *
- *   THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- *   IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- *   FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- *   AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- *   LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- *   OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- *   THE SOFTWARE.
- */
-package org.jeasy.random;
-
-import static org.junit.jupiter.api.Assertions.assertNotNull;
-import static org.junit.jupiter.api.Assertions.assertTrue;
-
-import org.junit.jupiter.api.Test;
-
-public class EasyRandomChainedSettersTest {
-
-    @Test
-    void chainSettersShouldBeRandomized() {
-        EasyRandom easyRandom = new EasyRandom(new EasyRandomParameters());
-
-        SubClassA value = easyRandom.nextObject(SubClassA.class);
-
-        assertNotNull(value.getField());
-        assertTrue(value.getF() != 0);
-    }
-
-    public static class BaseClass {
-
-        // setter overriden
-        private String field;
-        // single-char name
-        private int f;
-
-        public BaseClass setField(String field) {
-            this.field = field;
-            return this;
-        }
-
-        public String getField() {
-            return field;
-        }
-
-        public int getF() {
-            return f;
-        }
-
-        public BaseClass setF(int f) {
-            this.f = f;
-            return this;
-        }
-    }
-
-    public static class SubClassA extends BaseClass {
-
-        private SubClassB subClassB;
-    }
-
-    public static class SubClassB extends BaseClass {
-
-        @Override
-        public SubClassB setField(String field) {
-            super.setField(field);
-            return this;
-        }
-    }
-}
