--- easy-random-core/src/test/java/org/jeasy/random/EasyRandomOverridenFieldTest.java
@@ -1,59 +1,0 @@
-/*
- * The MIT License
- *
- *   Copyright (c) 2020, Mahmoud Ben Hassine (mahmoud.benhassine@icloud.com)
- *
- *   Permission is hereby granted, free of charge, to any person obtaining a copy
- *   of this software and associated documentation files (the "Software"), to deal
- *   in the Software without restriction, including without limitation the rights
- *   to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- *   copies of the Software, and to permit persons to whom the Software is
- *   furnished to do so, subject to the following conditions:
- *
- *   The above copyright notice and this permission notice shall be included in
- *   all copies or substantial portions of the Software.
- *
- *   THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- *   IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- *   FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- *   AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- *   LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- *   OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- *   THE SOFTWARE.
- */
-package org.jeasy.random;
-
-import static org.junit.jupiter.api.Assertions.assertFalse;
-import static org.junit.jupiter.api.Assertions.assertTrue;
-
-import org.junit.jupiter.api.Test;
-
-public class EasyRandomOverridenFieldTest {
-
-    @Test
-    void chainSettersShouldBeRandomized() {
-        EasyRandom easyRandom = new EasyRandom(new EasyRandomParameters());
-
-        SubClass value = easyRandom.nextObject(SubClass.class);
-
-        assertFalse(((BaseClass) value).field.isEmpty());
-        assertTrue(value.field != 0);
-    }
-
-    public static class BaseClass {
-
-        // field overriden (note: not setter)
-        private String field;
-
-        public BaseClass setField(String field) {
-            this.field = field;
-            return this;
-        }
-    }
-
-    public static class SubClass extends BaseClass {
-
-        // note: field overriden
-        private int field;
-    }
-}
